Replaced unsigned int with size_t

+++ src/beachline.cpp
@@ -421,10 +421,10 @@ void BeachLine<O>::insertAfter(SkipNode<O>* node, SkipNode<O>* at)
 
 inline int log2(int n)
 {
-    const unsigned int b[] = {0x2, 0xC, 0xF0, 0xFF00, 0xFFFF0000};
-    const unsigned int S[] = {1, 2, 4, 8, 16};
+    const size_t b[] = {0x2, 0xC, 0xF0, 0xFF00, 0xFFFF0000};
+    const size_t S[] = {1, 2, 4, 8, 16};
 
-    unsigned int r = 0; // result of log2(v) will go here
+    size_t r = 0; // result of log2(v) will go here
     for (int i = 3; i >= 0; i--) // unroll for speed (if compiler doesn't already)
     {
       if (n & b[i])
